séparation des fonctions dans le main dans un autre fichier + gestion des liaisons entre enquetes
