+++ fichiers/enquetes.json
@@ -1,119 +1,39 @@
 [
-  {
-    "id": 1,
-    "nom": "Vol",
-    "date_de_debut": "01/01/2023",
-    "date_de_fin": "30/06/2023",
-    "liste_preuves": [],
-    "personne_impliquee": [],
-    "liste_evenement": [],
-    "enquetes_liees": []
-  },
-  {
-    "id": 4,
-    "nom": "Cambriolage",
-    "date_de_debut": "15/06/2010",
-    "date_de_fin": "01/08/2011",
-    "liste_preuves": [],
-    "personne_impliquee": [
-      {
-        "nom": "Demarcq",
-        "prenom": "Alexis",
-        "date_de_naissance": "2003-08-04",
-        "date_de_deces": "9999-12-31",
-        "sexe": "Homme",
-        "metier": "'Pas de métier actuellement'",
-        "interrogatoires": {},
-        "mail": ""
-      },
-      {
-        "nom": "Lemaire",
-        "prenom": "Nathan",
-        "date_de_naissance": "2003-01-01",
-        "date_de_deces": "9999-12-31",
-        "sexe": "Homme",
-        "metier": "'Pas de métier actuellement'",
-        "interrogatoires": {},
-        "mail": ""
-      }
-    ],
-    "liste_evenement": [],
-    "enquetes_liees": []
-  },
-  {
-    "id": "3857c749-edec-4239-a2d2-adff902ceaee",
-    "nom": "Caca",
-    "date_de_debut": "14/12/2024",
-    "date_de_fin": "14/12/2025",
-    "liste_preuves": [],
-    "personne_impliquee": [],
-    "liste_evenement": [],
-    "enquetes_liees": []
-  },
-  {
-    "id": "84bada52-4b02-4456-a879-6e4085cb28d5",
-    "nom": "Meurtre",
-    "date_de_debut": "04/08/2003",
-    "date_de_fin": "26/02/2005",
-    "liste_preuves": [],
-    "personne_impliquee": [
-      {
-        "nom": "Demarcq",
-        "prenom": "Alexis",
-        "date_de_naissance": "2003-08-04",
-        "date_de_deces": "9999-12-31",
-        "sexe": "Homme",
-        "metier": "'Pas de métier actuellement'",
-        "interrogatoires": {},
-        "mail": ""
-      },
-      {
-        "nom": "Henrard",
-        "prenom": "Quentin",
-        "date_de_naissance": "2003-08-04",
-        "date_de_deces": "9999-12-31",
-        "sexe": "Homme",
-        "metier": "'Pas de métier actuellement'",
-        "interrogatoires": {},
-        "mail": ""
-      },
-      {
-        "nom": "Mertens",
-        "prenom": "Corentin",
-        "classe": "Temoin",
-        "sexe": "Homme",
-        "date_de_naissance": "2003-09-20",
-        "metier": "'Pas de métier actuellement'",
-        "mail": "",
-        "interrogatoires": {},
-        "temoignage": {
-          "2024-12-14": ["J'étais en train de travailler sur mon projet d'OS2"]
-        },
-        "fiabilite": 0,
-        "protection": false,
-        "disponibilite": false
-      },
-      {
-        "nom": "Henrard",
-        "prenom": "Quentin",
-        "classe": "Criminel",
-        "sexe": "Homme",
-        "date_de_naissance": "2003-08-04",
-        "metier": "'Pas de métier actuellement'",
-        "mail": "",
-        "interrogatoires": {},
-        "innocent": false,
-        "statut_legal": false,
-        "suspection": "",
-        "alibi": "",
-        "apparence": {},
-        "corpulence": {},
-        "dossier_psychologique": {},
-        "lieu_de_detention": "",
-        "niveau_de_dangerosite": 1
-      }
-    ],
-    "liste_evenement": [],
-    "enquetes_liees": []
-  }
+    {
+        "id": "f0043918-317f-44f9-9bf3-c7e73b8b53a2",
+        "nom": "meurtre",
+        "date_de_debut": "28/06/2005",
+        "date_de_fin": "28/06/2025",
+        "liste_preuves": [],
+        "personne_impliquee": [
+            {
+                "nom": "Demarcq",
+                "prenom": "Alexis",
+                "classe": "Suspect",
+                "sexe": "Homme",
+                "date_de_naissance": "2003-08-04",
+                "metier": "'Pas de métier actuellement'",
+                "mail": "",
+                "interrogatoires": {},
+                "innocent": false,
+                "statut_legal": false,
+                "suspection": "",
+                "alibi": ""
+            }
+        ],
+        "liste_evenement": [],
+        "enquetes_liees": [
+            "5302816f-cabd-4771-a565-8ed3bf0c5db4"
+        ]
+    },
+    {
+        "id": "fddc9710-62c2-442a-839c-f29c0cdab8e6",
+        "nom": "trolley",
+        "date_de_debut": "13/04/1546",
+        "date_de_fin": "31/12/9999",
+        "liste_preuves": [],
+        "personne_impliquee": [],
+        "liste_evenement": [],
+        "enquetes_liees": []
+    }
 ]
