+++ fichiers/personnes.json
@@ -7,7 +7,11 @@
         "date_de_naissance": "2003-08-04",
         "metier": "'Pas de métier actuellement'",
         "mail": "",
-        "interrogatoires": {}
+        "interrogatoires": {},
+        "innocent": false,
+        "statut_legal": false,
+        "suspection": "",
+        "alibi": ""
     },
     {
         "nom": "Henrard",
